Add files via upload
For task 1 given the expected returns as μ = (μ1,------- μn ) , correlation matrix C and a given expected return of the portfolio, r = μT x, using Markowitz portfolio optimization model , we have to determine the vector x = (x1, ------ xn )T that minimize the variance of portfolio. Also, 25 different values of ‘r’ are given to us as –
r = 3.00 ,3.25 ,3.50 ,3.75 ,-------, 8.50, 8.75, 9.00
Now the implementation for the above problem is done in python using solver ‘Gurobipy’

The portfolio optimization problem solved in Task 1 assumed the entire capital is invested and in code we used sum of portfolio weights as 1. However, in Task 2 there is some relaxation provided and allowed to keep a fraction of capital uninvested without any return or risk. To get this a small change in constraint is done which is as follows:
𝒆𝑇𝒙 = 1 this constraint is used in Task 1 which is changed to 𝒆𝑇𝒙 ≤ 1 for Task 2.
This depicts a scenario where an investor can decide to save a portion of portfolio without any return or risk.

Constraint for Expected Returns defined in Task 1 is ‘𝝁𝑇𝒙 = 𝑟’ , which means that the expected return of the portfolio is same or equals to the target return. Now for Task 3 this constraint is modified to ‘𝝁𝑇𝒙 >= 𝑟’ , that shows the expected return of the portfolio is now greater than or equal to target returns.
Accordingly, the Markowitz Portfolio Optimization function is modified in the code and the constraint for Expected Return is changed .

Task 4
Short Selling
Short Selling in finance which is also known as “shorting” or “selling short” implies the sale of a financial security that is borrowed by the seller. It refers to the ability to sell security or assets that are not owned currently with the intention of buying them back in future. The assets are borrowed and sold at the current market price and repurchased later to return to the lender. The seller believes that the borrowed security’s price will decrease, which in turn enables it to be bought back at a lower price for profit.
To model Short Selling in portfolio optimization, the constraint which enforces the non-negativity of the portfolio weights, is removed. Now for some assets there exist non-negative weights, indicating that assets are selling short.